working Character and Camera
The character now moves in all directions and jumps, the camera has
various settings to tweak the way it's looking, height and distance. An
important object is the GameControl that receives all the input from
the keyboard and mouse, and then distributes the comands via a system
of subscribers. In this case the only subscribers are the camera and
the player. But any gameObject can subscribe since the function for it
is static. This makes it easy.
Also the camera only takes in one external variable, the target, this
makes it easy for it to switch characters to follow.
You can also toggle 3rd person view and 1rst person view by pressing
"p".
To apply these scripts to a project just make sure you have a
GameControl (empty gameObject) for the input and remember to assign a
target to the camera. Apart form that it's just click and drag.